Refactor: Moved widget for selecting/deleting a file and the image preview to separate components
