+++ assets/pages/Recipes/constants/DATEI_AUSWAEHLEN.ts
@@ -1,0 +1,1 @@
+export const DATEI_AUSWAEHLEN = 'Datei auswählen'
